+++ docs/kernel/wave_codegen.rst
@@ -1,6 +1,6 @@
 Wave Codegen
 ============
 
-.. automodule:: wave_lang.kernel.wave.codegen
+.. automodule:: wave_lang.kernel.compiler.wave_codegen
   :members:
   :undoc-members:
